requirements.txt
asgiref==3.7.2
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==41.0.5
Django==4.2.7
django-livesync==0.5
idna==3.4
mysqlclient==2.2.0
pycparser==2.21
PyJWT==2.4.0
pytz==2023.3.post1
requests==2.31.0
sqlparse==0.4.4
tornado==6.3.3
typing_extensions==4.8.0
urllib3==2.1.0
watchdog==3.0.0
websocket-client==1.6.4
